Meu_MVP/requirements.txt
pandas
scikit-learn
matplotlib
numpy
seaborn
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic